Ditch middleware in favor of custom authentication backend
