+++ normandy/base/tests/test_middleware.py
@@ -1,17 +1,11 @@
-import json
 from random import randint
 
 import pytest
-from requests.exceptions import ConnectionError
-from django.core.exceptions import PermissionDenied
-from django.contrib.auth.models import User
 
 from normandy.base.middleware import (
     NormandyCommonMiddleware,
     NormandySecurityMiddleware,
     DEBUG_HTTP_TO_HTTPS_REDIRECT,
-    OIDCAccessTokenAuthorizationMiddleware,
-    OIDCEndpointRequestError,
 )
 
 
@@ -93,176 +87,3 @@ class TestNormandySecurityMiddleware(object):
                 },
             }
         )
-
-
-class TestOIDCAccessTokenAuthenticationMiddleware:
-
-    @pytest.fixture
-    def middleware(self):
-        return OIDCAccessTokenAuthorizationMiddleware(
-            lambda req: 'response'
-        )
-
-    @pytest.mark.django_db
-    def test_happy_path(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        user_profile = {
-            'email': 'Peterbe@example.com ',  # note the trailing space
-            'given_name': 'Peter ',
-            'family_name': 'Bengtsson',
-        }
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            content=json.dumps(user_profile).encode('utf-8'),
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer perfectly-valid')
-        res = middleware(req)
-        assert res == 'response'  # Meaning, it didn't interfer the response
-        # This should have created a user
-        user = User.objects.get()
-        # The middleware sets `request.user = found_user`
-        assert req.user
-        user = req.user
-        assert user.email == 'peterbe@example.com'
-        assert user.username == 'peterbe@example.com'
-        assert user.first_name == 'Peter'
-        assert user.last_name == 'Bengtsson'
-        assert not user.is_staff
-        assert not user.is_superuser
-        # Also, sanity check that the user got created
-        assert User.objects.get().username == 'peterbe@example.com'
-
-    @pytest.mark.django_db
-    def test_happy_path_user_already_exists(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        existing_user = User.objects.create(
-            username='anything',
-            email='peterbe@example.com'
-        )
-        assert not existing_user.last_login
-        user_profile = {
-            'email': 'Peterbe@example.com ',  # note the trailing space
-            'given_name': 'Peter ',
-            'family_name': 'Bengtsson',
-        }
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            content=json.dumps(user_profile).encode('utf-8'),
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer perfectly-valid')
-        middleware(req)
-        existing_user.refresh_from_db()
-        assert existing_user.last_login
-
-    def test_nothing_if_no_header(self, middleware, rf):
-        req = rf.get('/')
-        res = middleware(req)
-        # This means, nothing happened in the middleware's process_request()
-        assert res == 'response'
-        # Same if it's set but empty
-        req = rf.get('/', HTTP_AUTHORIZATION='')
-        res = middleware(req)
-        assert res == 'response'
-
-    def test_bad_bearer_token(self, middleware, rf):
-        req = rf.get('/', HTTP_AUTHORIZATION='Any old junk')
-        with pytest.raises(PermissionDenied) as err:
-            middleware(req)
-        assert 'Invalid Authorization header value' in str(err)
-
-    def test_bad_access_token(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            status_code=401,
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer notgoodenough')
-        with pytest.raises(PermissionDenied) as err:
-            middleware(req)
-        assert 'Unauthorized access token' in str(err)
-
-    def test_oidc_provider_unhappy(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        """If the communication between our server and the OIDC provider works but
-        the response is a 4xx error (but not 401) then an error will be raise."""
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            status_code=400,
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer perfectly-valid')
-        with pytest.raises(OIDCEndpointRequestError) as err:
-            middleware(req)
-        assert '400' in str(err), str(err)
-
-    @pytest.mark.django_db
-    def test_oidc_provider_retries_on_5xx_errors(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        # This will cause a retry
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            status_code=504,
-        )
-        user_profile = {
-            'email': 'Peterbe@example.com',
-            'given_name': 'Peter ',
-            'family_name': 'Bengtsson',
-        }
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            content=json.dumps(user_profile).encode('utf-8'),
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer notgoodenough')
-        middleware(req)
-        # The middleware sets `request.user = found_user`
-        assert req.user
-
-    @pytest.mark.django_db
-    def test_oidc_provider_retries_on_request_connection_errors(
-        self,
-        middleware,
-        rf,
-        requestsmock,
-        settings,
-    ):
-        # This will cause a retry
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            exc=ConnectionError,
-        )
-        user_profile = {
-            'email': 'Peterbe@example.com',
-            'given_name': 'Peter ',
-            'family_name': 'Bengtsson',
-        }
-        requestsmock.get(
-            settings.OIDC_USER_ENDPOINT,
-            content=json.dumps(user_profile).encode('utf-8'),
-        )
-        req = rf.get('/', HTTP_AUTHORIZATION='Bearer notgoodenough')
-        middleware(req)
-        assert req.user
